refactoring of api access for dfu modal
